fix(lv_pos): fix invalidate area problem with transform
Revert "fix(lv_pos): fix invalidate area problem with transform"

This reverts commit ced575402046fb0959396fc761434823a442f86a.

fix sin table precise and img buf transform area


remove debug logs


add lv_obj_invalidate in add/remove/set local style 


add LV_STYLE_PROP_FLAG_TRANSFORM

+++ src/misc/lv_math.h
@@ -44,9 +44,9 @@ typedef struct {
  * @param angle
  * @return sinus of 'angle'. sin(-90) = -32767, sin(90) = 32767
  */
-LV_ATTRIBUTE_FAST_MEM int16_t lv_trigo_sin(int16_t angle);
+LV_ATTRIBUTE_FAST_MEM int32_t lv_trigo_sin(int16_t angle);
 
-static inline LV_ATTRIBUTE_FAST_MEM int16_t lv_trigo_cos(int16_t angle)
+static inline LV_ATTRIBUTE_FAST_MEM int32_t lv_trigo_cos(int16_t angle)
 {
     return lv_trigo_sin(angle + 90);
 }
